# cmd/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Decentralized Payroll and Invoicing Platform for Remote Teams",
        "title": "DefiFundr API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DefiFundr Support",
            "url": "http://defifundr.com/support",
            "email": "hello@defifundr.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/waitlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List waitlist entries with pagination and filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "waitlist"
                ],
                "summary": "List waitlist entries",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (waiting, invited, registered)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by referral source",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by (signup_date_asc, signup_date_desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of waitlist entries",
                        "schema": {
                            "$ref": "#/definitions/response.PageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/waitlist/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Export all waitlist data as CSV",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "waitlist"
                ],
                "summary": "Export waitlist data",
                "responses": {
                    "200": {
                        "description": "CSV file",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/admin/waitlist/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get statistics about the waitlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "waitlist"
                ],
                "summary": "Get waitlist statistics",
                "responses": {
                    "200": {
                        "description": "Waitlist statistics",
                        "schema": {
                            "$ref": "#/definitions/response.WaitlistStatsResponse"
                        }
                    }
                }
            }
        },
...
        "/waitlist": {
            "post": {
                "description": "Register for early access to the platform",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "waitlist"
                ],
                "summary": "Join the waitlist",
                "parameters": [
                    {
                        "description": "Waitlist join data",
                        "name": "join",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.WaitlistJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully joined waitlist",
                        "schema": {
                            "$ref": "#/definitions/response.WaitlistEntryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email already on waitlist",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.WaitlistJoinRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "referral_code": {
                    "type": "string"
                },
                "referral_source": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.PageResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "response.WaitlistEntryResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_date": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "referral_code": {
                    "type": "string"
                },
                "referral_source": {
                    "type": "string"
                },
                "signup_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.WaitlistStatsResponse": {
            "type": "object",
            "properties": {
                "conversion_rate": {
                    "type": "number"
                },
                "invited_count": {
                    "type": "integer"
                },
                "registered_count": {
                    "type": "integer"
                },
                "sources": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "total_signups": {
                    "type": "integer"
                },
                "waiting_count": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}